[Version 3.1.7]
1) Changed video stream player from VLC to ExoPlayer (by Google) that is optimised for Android; Added a developing (unofficial) version for the available use of its RTSP streaming feature, that is otherwise unavailable in its current official version.

2) Altered video stream fragment with the following modification:
   i) CCT now has 6 video streams that can be run concurrently and each one can be displayed in full screen.
   ii) There is now a home stream which is represented by a home icon. This is used for the user's own video stream.
   iii) Video links will be played automatically upon the save action in the order that the streams are added; i.e. home stream plays in the 1st video, subsequent video added will play in the video panel of 2nd, 3rd etc.
   iv) This applies to the video stream for dashboard as well.

3) Refined Wave Relay code to successfully detect for USB connection. This is used for the radio link status of the user.

4) Implemented HD photos compression (captured or chosen) to be small enough for storage/transfer. This does not reduce the quality of the images.

5) Added BW Paris blueprints and measurement for BFT. (Remaining issue now is how the 'factor' value is derived, which is not what was explained.)

6) Added BFT modules directly into Ventilo app. This is so that gradle does not have to access Kah Siong's home server for updates as there is no further updates to be done. Previously, when gradle tries to access the home server, it will not be able to compile if it is switched off.

7) Bugs Resolved
    i) Own blueprint location not saved when navigate to other fragments and back
    ii) Hazard and Deceased Icons appear in almost every deck
    iii) Sit Rep image not saved
    iv) Fragment title disappears and icon selection resets to Map Blueprint when app goes to sleep (onCreate)
    v) Added Network Stop Service to ensure all services have been stopped upon closing of app by user

+++ app/src/main/java/sg/gov/dsta/mobileC3/ventilo/network/jeroMQ/JeroMQPublisher.java
@@ -21,14 +21,17 @@ public class JeroMQPublisher extends JeroMQParentPublisher {
 
     //    private static final Logger LOGGER = LoggerFactory.getLogger(JeroMQPublisher.class);
     private static final String TAG = JeroMQPublisher.class.getSimpleName();
-//    private static final int DEFAULT_PORT = 5556;
+    private static final int DEFAULT_PORT = 5556;
 
 //    protected static final String SERVER_PUB_IP_ADDRESS = "tcp://" +
 //            NetworkUtil.getOwnIPAddressThroughWiFiOrEthernet(true) + ":" +
 //            JeroMQPubSubBrokerProxy.DEFAULT_XSUB_PORT;
 
-    protected static final String SERVER_PUB_IP_ADDRESS = "tcp://*:" +
-            JeroMQPubSubBrokerProxy.DEFAULT_XSUB_PORT;
+//    protected static final String SERVER_PUB_IP_ADDRESS = "tcp://" +
+//            NetworkUtil.getOwnIPAddressThroughWiFiOrEthernet(true) + ":" +
+//            DEFAULT_PORT;
+
+    protected static final String SERVER_PUB_IP_ADDRESS = "tcp://*:" + DEFAULT_PORT;
 
 //    protected static final String SERVER_PUB_IP_ADDRESS = "tcp://192.168.1.3:" +
 //            JeroMQPubSubBrokerProxy.DEFAULT_XSUB_PORT;
@@ -140,6 +143,8 @@ public class JeroMQPublisher extends JeroMQParentPublisher {
         mZContext = new ZContext();
         mPubSocket = mZContext.createSocket(SocketType.PUB);
         mPubSocket.setMaxMsgSize(-1);
+//        mPubSocket.setMsgAllocationHeapThreshold(1024 * 1024);
+//        mPubSocket.setSendBufferSize(1024 * 1024);
         mPubSocket.setLinger(0);
 //        mPubSocket.connect(SERVER_PUB_IP_ADDRESS);
         mPubSocket.bind(SERVER_PUB_IP_ADDRESS);
@@ -459,8 +464,16 @@ public class JeroMQPublisher extends JeroMQParentPublisher {
 //        task.execute();
 
         if (mPubSocket != null) {
-            mZContext.destroySocket(mPubSocket);
+            mPubSocket.unbind(SERVER_PUB_IP_ADDRESS);
+            Log.i(TAG, "Server pub socket unbound.");
+
+            mPubSocket.close();
             Log.i(TAG, "Server pub socket closed.");
+
+            if (mZContext != null) {
+                mZContext.destroySocket(mPubSocket);
+                Log.i(TAG, "Server pub socket destroyed.");
+            }
         }
 
         if (mZContext != null && !mZContext.isClosed()) {
@@ -468,5 +481,7 @@ public class JeroMQPublisher extends JeroMQParentPublisher {
             mZContext.destroy();
             Log.i(TAG, "ZContext destroyed.");
         }
+
+
     }
 }
